lab5_evaluation/pyproject.toml
[project]
name = "lab5-evaluation"
version = "0.1.0"
description = "Lab 5: Agent Evaluation & Testing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[eval]>=1.19.0",
    "python-dotenv>=1.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
